バグ取り切ったよ
ほめて

↓↓解決したバグ
弓矢のダメージができてない
ジャンプエフェクト消して
狂戦士ノックバックしすぎ
連打で2倍ノックバック
ノックバックしながらスタン
重鎧兵か田杉
重鎧兵の空中を右クリックしても反応してる
勇者がワンパンマン
ゲーム終了時にHPを回復

+++ src/main/java/jp/meao0525/battleemblem/battleclass/ClassStatus.java
@@ -1,12 +1,12 @@
 package jp.meao0525.battleemblem.battleclass;
 
 public enum ClassStatus {
-    SWORD_MASTER_STATUS(40.0,0.3F,6,0),
-    BERSERKER_STATUS(40.0,0.3F,8,0),
-    ARMOR_KNIGHT_STATUS(60.0,0.2F,6,2),
-    BRAVE_HERO_STATUS(40.0,0.3F,4,1),
-    SNIPER_STATUS(40.0,0.3F,1,0),
-    ASSASSIN_STATUS(40.0,0.3F,4,0);
+    SWORD_MASTER_STATUS(40.0,0.3F,6.0,0.0),
+    BERSERKER_STATUS(40.0,0.3F,8.0,0.0),
+    ARMOR_KNIGHT_STATUS(60.0,0.2F,6.0,2.0),
+    BRAVE_HERO_STATUS(40.0,0.3F,4.0,1.0),
+    SNIPER_STATUS(40.0,0.3F,1.0,0.0),
+    ASSASSIN_STATUS(40.0,0.3F,4.0,0.0);
 
 
     private final double hp;
@@ -14,7 +14,7 @@ public enum ClassStatus {
     private final double attack;
     private final double defence;
 
-    private ClassStatus(double hp, float speed, int attack, int defence) {
+    private ClassStatus(double hp, float speed, double attack, double defence) {
         this.hp = hp;
         this.speed = speed;
         this.attack = attack;
